Rename ordinalBand to band, ordinalPoint to point.
