+++ index.js
@@ -1,9 +1,9 @@
+export {default as band} from "./src/band";
 export {default as identity} from "./src/identity";
 export {default as linear} from "./src/linear";
 export {default as log} from "./src/log";
 export {default as ordinal, reuse as ordinalReuse} from "./src/ordinal";
-export {default as ordinalBand} from "./src/ordinalBand";
-export {default as ordinalPoint} from "./src/ordinalPoint";
+export {default as point} from "./src/point";
 export {default as pow, sqrt} from "./src/pow";
 export {default as quantile} from "./src/quantile";
 export {default as quantize} from "./src/quantize";
