+++ src/band.js
