+++ src/point.js
