+++ test/band-test.js
