+++ test/point-test.js
